Add an Estimate model and let it tell us about itself
I've moved all the estimate related code to a PORO Estimate. It knows
how to format itself. I don't know about sticking the CSS classes in
there, but it "works".

The goal was to use a standard Rails form and "decorate" with the nice
estimate control. The JS controller now just reads/writes from the
hidden Rails select.